% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{combine_sources}
\alias{combine_sources}
\title{Stack harmonized sources into one long dataset}
\usage{
combine_sources(...)
}
\arguments{
\item{...}{\code{ch_harmonized} tibbles (or a list of them).}
}
\description{
Stack harmonized sources into one long dataset
}
