% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{coverage_report}
\alias{coverage_report}
\title{Coverage report: how populated each harmonized variable is}
\usage{
coverage_report(harmonized, dataschema = attr(harmonized, "dataschema"))
}
\arguments{
\item{harmonized}{A \code{ch_harmonized} object.}

\item{dataschema}{Schema (defaults to the attached one).}
}
\value{
A tibble: variable, domain, n_populated, pct_populated, value_range.
}
\description{
Coverage report: how populated each harmonized variable is
}
