% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{select}
\alias{select}
\alias{select_domain}
\alias{select_variables}
\title{Pull a domain or specific variables from a harmonized dataset}
\usage{
select_domain(harmonized, domains, dataschema = attr(harmonized, "dataschema"))

select_variables(harmonized, vars)
}
\arguments{
\item{harmonized}{A \code{ch_harmonized} object.}

\item{domains, vars}{Character vectors.}

\item{dataschema}{Schema with a \code{domain} column (defaults to the attached one).}
}
\description{
Pull a domain or specific variables from a harmonized dataset
}
